% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvoCWGEE.R
\name{mvoCWGEE}
\alias{mvoCWGEE}
\title{Cluster weighted GEE for multiple correlated binary outcomes in cross-sectional data with informative cluster size.}
\usage{
mvoCWGEE(formula, data, cluster, resp.ind, unit, corr.str,
  common.slope = NULL)
}
\arguments{
\item{formula}{a formula expression as for other regression models.}

\item{data}{an optional data frame containing the variables provided in
\code{formula}, \code{id}, \code{cluster.var} and \code{time.var}.}

\item{cluster}{a vector that identifies the clusters.}

\item{resp.ind}{a vector that indicates the responses.}

\item{unit}{a vector that identifies the unit within a cluster.}

\item{corr.str}{a character string that indicates the working correlation structure among the correlated responses.
Options include \code{"ind"} for independence, \code{"unstr"} for unstructured, and \code{"exch"} for exchangeable.}

\item{common.slope}{a character string indicating which variables in the model will
have a common slope for each of the responses.}
}
\value{
Returns an object of the class \code{"cwgee"}. This has components:
\item{call}{the matched call.}
\item{coefficients}{the estimated regression parameter vector of the marginal model.}
\item{coef.names}{the variable name of the coefficients.}
\item{robust.variance}{the estimated "robust" covariance matrix.}
\item{robust.se}{the estimated "robust" standard errors.}
\item{wald.chisq}{the Wald Chi-square test statistic for coefficient estimates.}
\item{p.value}{the p-value based on a Wald Chi-square test statistic that no covariates are statistically significant.}
\item{corr.matrix}{the estimated correlation matrix.}
\item{niter}{the number of iterations the model took to converge.}
\item{corr.str}{the working correlation structure assumed for the model.}
}
\description{
Solves the cluster-weighted generalized estimating equations for correlated binary
responses in clustered data assuming
using the method of quasi-least squares.
}
\details{
The \code{data} must be provided in case level or equivalently in `long' format.
}
\examples{
data(perio_base)
fitmod <- mvoCWGEE(formula = y ~ smoking + age + edu, data = perio_base,
cluster = subject, resp.ind = outcome, unit = tooth,
common.slope = c("smoking", "edu"), corr.str = "exch")
summary(fitmod)
}
\author{
Aya Mitani
}
