% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordCWGEE.R
\name{ordCWGEE}
\alias{ordCWGEE}
\title{Cluster weighted GEE for ordinal clustered longitudinal data with informative cluster size.}
\usage{
ordCWGEE(formula, data, id, cluster.var, time.var, time.str)
}
\arguments{
\item{formula}{a formula expression as for other regression models.}

\item{data}{an optional data frame containing the variables provided in
\code{formula}, \code{id}, \code{cluster.var} and \code{time.var}.}

\item{id}{a vector that identifies the clusters.}

\item{cluster.var}{a vector that identifies the unit within a cluster.}

\item{time.var}{a vector that identifies the repeated observation of a unit.}

\item{time.str}{a character string that indicates the temporal working correlation structure.
Options include \code{"ind"} for independence, \code{"ar1"} for AR1, and \code{"exch"} for exchangeable.}
}
\value{
Returns an object of the class \code{"cwgee"}. This has components:
\item{call}{the matched call.}
\item{coefficients}{the estimated regression parameter vector of the marginal model.}
\item{coef.names}{the variable name of the coefficients.}
\item{robust.variance}{the estimated "robust" covariance matrix.}
\item{robust.se}{the estimated "robust" standard errors.}
\item{wald.chisq}{the Wald Chi-square test statistic for coefficient estimates.}
\item{p.value}{the p-value based on a Wald Chi-square test statistic that no covariates are statistically significant.}
\item{alpha}{the estimated temporal correlation coefficient.}
\item{niter}{the number of iterations the model took to converge.}
\item{time.str}{the temporal working correlation structure assumed for the model.}
}
\description{
Solves the cluster-weighted generalized estimating equations for correlated ordinal
responses in clustered longitudinal data assuming a
cumulative link logit model for the marginal probabilities
using the method of quasi-least squares.
}
\details{
The \code{data} must be provided in case level or equivalently in `long' format.
}
\examples{
data(perio)
fitmod <- ordCWGEE(formula = cal ~ mets + edu + age + smoking, data = perio,
id = subject, cluster.var = tooth, time.var = visit, time.str = "ind")
summary(fitmod)
}
\author{
Aya Mitani
}
